23 commit admin commands cleaned

+++ src/main/java/by/ksu/training/controller/commands/admin/DeleteAssignedTrainerCommand.java
@@ -16,26 +16,40 @@ import javax.servlet.http.HttpServletResponse;
 import java.util.List;
 
 /**
+ * Deletes records of AssignedTrainer.
+ *
  * @Author Kseniya Oznobishina
  * @Date 20.01.2021
  */
 public class DeleteAssignedTrainerCommand extends AdminCommand {
     private static Logger logger = LogManager.getLogger(DeleteAssignedTrainerCommand.class);
 
+    /**
+     * Deletes records of AssignedTrainer with id which came in parameter "remove".
+     * @return ResponseState
+     * @throws PersistentException if any exception occur in service layout.
+     * @see by.ksu.training.controller.state.ResponseState
+     * @see by.ksu.training.entity.AssignedTrainer
+     */
     @Override
     protected ResponseState exec(HttpServletRequest request, HttpServletResponse response) throws PersistentException {
-            Validator<AssignedTrainer> validator = new AssignedTrainerValidator();
-            List<Integer> listId = validator.validateListId(AttrName.REMOVE, request);
+        Validator<AssignedTrainer> validator = new AssignedTrainerValidator();
+        AssignedTrainerService atService = factory.getService(AssignedTrainerService.class);
 
-            AssignedTrainerService atService = factory.getService(AssignedTrainerService.class);
+        String action = request.getParameter(AttrName.ACTION);
+        String parameter = action == null ? "" : "?" + AttrName.ACTION + "=" + action;
+        ResponseState state = new RedirectState("assigned_trainer/list.html" + parameter);
+
+        List<Integer> listId = validator.validateListId(AttrName.REMOVE, request);
+        if (!listId.isEmpty()) {
             for (Integer id : listId) {
                 atService.delete(id);
+                logger.debug("Admin deleted AssignedTrainer with id={}", id);
             }
-
-            request.getSession().setAttribute(AttrName.SUCCESS_MESSAGE, "message.success.delete");
-            String action = request.getParameter(AttrName.ACTION);
-            String parameter = action == null ? "" : "?" + AttrName.ACTION + "=" + action;
-            return new RedirectState("assigned_trainer/list.html" + parameter);
-
+            state.getAttributes().put(AttrName.SUCCESS_MESSAGE, "message.success.delete");
+        } else {
+            state.getAttributes().put(AttrName.WARNING_MAP, validator.getWarningMap());
+        }
+        return state;
     }
 }
