+++ src/main/java/by/ksu/training/controller/commands/admin/DeleteExerciseCommand.java
@@ -16,24 +16,38 @@ import javax.servlet.http.HttpServletResponse;
 import java.util.List;
 
 /**
+ * Deletes records of Exercises.
+ *
  * @Author Kseniya Oznobishina
  * @Date 24.01.2021
  */
-public class DeleteExerciseCommand extends AdminCommand{
+public class DeleteExerciseCommand extends AdminCommand {
     private static Logger logger = LogManager.getLogger(DeleteExerciseCommand.class);
+
+    /**
+     * Deletes records of Exercises with id which came in parameter "remove".
+     *
+     * @return ResponseState
+     * @throws PersistentException if any exception occur in service layout.
+     * @see by.ksu.training.controller.state.ResponseState
+     */
 
     @Override
     protected ResponseState exec(HttpServletRequest request, HttpServletResponse response) throws PersistentException {
-            Validator<Exercise> validator = new ExerciseValidator();
-            List<Integer> listId = validator.validateListId(AttrName.REMOVE, request);
-            ExerciseService service = factory.getService(ExerciseService.class);
+        Validator<Exercise> validator = new ExerciseValidator();
+        ExerciseService service = factory.getService(ExerciseService.class);
+        ResponseState state = new RedirectState("exercise/list.html");
+
+        List<Integer> listId = validator.validateListId(AttrName.REMOVE, request);
+        if (!listId.isEmpty()) {
             for (Integer id : listId) {
                 service.delete(id);
+                logger.debug("Admin deleted Exercise with id={}", id);
             }
-
-            request.getSession().setAttribute(AttrName.SUCCESS_MESSAGE, "message.success.delete");
-            return new RedirectState("exercise/list.html" );
-
-
+            state.getAttributes().put(AttrName.SUCCESS_MESSAGE, "message.success.delete");
+        } else {
+            state.getAttributes().put(AttrName.WARNING_MAP, validator.getWarningMap());
+        }
+        return state;
     }
 }
