+++ src/main/java/by/ksu/training/controller/commands/trainer/DeleteAssignedComplexCommand.java
@@ -4,8 +4,11 @@ import by.ksu.training.controller.AttrName;
 import by.ksu.training.controller.state.RedirectState;
 import by.ksu.training.controller.state.ResponseState;
 import by.ksu.training.entity.AssignedComplex;
+import by.ksu.training.entity.AssignedTrainer;
+import by.ksu.training.entity.User;
 import by.ksu.training.exception.PersistentException;
 import by.ksu.training.service.AssignedComplexService;
+import by.ksu.training.service.AssignedTrainerService;
 import by.ksu.training.service.validator.AssignedComplexValidator;
 import by.ksu.training.service.validator.Validator;
 import org.apache.logging.log4j.LogManager;
@@ -16,26 +19,50 @@ import javax.servlet.http.HttpServletResponse;
 import java.util.List;
 
 /**
+ * Deletes records of AssignedComplex.
+ *
  * @Author Kseniya Oznobishina
  * @Date 23.01.2021
  */
+
 public class DeleteAssignedComplexCommand extends TrainerCommand {
     private static Logger logger = LogManager.getLogger(DeleteAssignedComplexCommand.class);
 
+    /**
+     * Deletes records of AssignedComplex with id which came in parameter "remove".
+     *
+     * @return ResponseState
+     * @throws PersistentException if any exception occur in service layout.
+     * @see by.ksu.training.controller.state.ResponseState
+     */
+
     @Override
     protected ResponseState exec(HttpServletRequest request, HttpServletResponse response) throws PersistentException {
-            Validator<AssignedComplex> validator = new AssignedComplexValidator();
+        Validator<AssignedComplex> validator = new AssignedComplexValidator();
+        AssignedComplexService service = factory.getService(AssignedComplexService.class);
+        AssignedTrainerService assignedTrainerService = factory.getService(AssignedTrainerService.class);
+
+        int visitorId = validator.validateIntAttr(AttrName.USER_ID, request);
+        String parameter = "?" + AttrName.USER_ID + "=" + visitorId;
+        ResponseState state = new RedirectState("assigned_complex/list.html" + parameter);
+        User visitor = new User(visitorId);
+        User trainer = (User) request.getSession().getAttribute(AttrName.AUTHORIZED_USER);
+        boolean allowed = assignedTrainerService.checkTrainerByVisitor(trainer, visitor);
+
+        if (allowed) {
             List<Integer> listId = validator.validateListId(AttrName.REMOVE, request);
-            AssignedComplexService service = factory.getService(AssignedComplexService.class);
-            for (Integer id : listId) {
-                service.delete(id);
+            if (!listId.isEmpty()) {
+                for (Integer id : listId) {
+                    service.delete(id);
+                    logger.debug("Trainer id={} deleted AssignedTrainer with id={}", trainer.getLogin(), id);
+                }
+                state.getAttributes().put(AttrName.SUCCESS_MESSAGE, "message.success.delete");
+            } else {
+                state.getAttributes().put(AttrName.WARNING_MAP, validator.getWarningMap());
             }
-
-            request.getSession().setAttribute(AttrName.SUCCESS_MESSAGE, "message.success.delete");
-            String userId = request.getParameter(AttrName.USER_ID);
-            String parameter = userId == null ? "" : "?" + AttrName.USER_ID + "=" + userId;
-            return new RedirectState("assigned_complex/list.html" + parameter);
-
-
+            return state;
+        } else {
+            throw new PersistentException(String.format("User %s are not allowed to delete AssignedComplex for visitor %s", trainer.getLogin(), visitor.getId()));
+        }
     }
 }
